\name{run_cellgpsr_app}
\alias{run_cellgpsr_app}
\title{Run the Cell-GPS Shiny heatmap app}
\usage{
run_cellgpsr_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::shinyApp}.}
}
\value{
A Shiny application object.
}
\description{
Starts the lightweight Cell-GPS Shiny app for loading a CSV matrix and exporting
a clustered heatmap PDF.
}
